Moved slsp related files
