+++ src/slsp/features/ct/CTResultDataProvider.ts
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 import { Event, EventEmitter, ProviderResult, TreeDataProvider, TreeItem, TreeItemCollapsibleState } from "vscode";
-import { CTResultPair } from "./slsp/protocol/combinatorialTesting";
+import { CTResultPair } from "../../protocol/combinatorialTesting";
 
 export class CTResultDataProvider implements TreeDataProvider<CTResultElement> {
